chore: add a check for an empty array before deleting

+++ playwright-tests/e2e/helpers/global-helpers.ts
@@ -82,15 +82,15 @@ export async function uploadFile({
 }
 
 export async function deleteFiles() {
-  try {
-    const filesResponse = (await axios.get(getStrapiUrl({ path: '/api/upload/files' }))).data;
+  const filesResponse = (await axios.get(getStrapiUrl({ path: '/api/upload/files' }))).data;
 
-    const filesDelete = filesResponse.filter((file) => file.name?.startsWith(E2E_SMOKE_NAME_PREFIX));
+  const filesDelete = filesResponse.filter((file) => file.name?.startsWith(E2E_SMOKE_NAME_PREFIX));
 
+  if (filesDelete.length) {
     filesDelete.forEach(async ({ id }) => {
       await axios.delete(getStrapiUrl({ path: `/api/upload/files/${id}` }));
     })
-  } catch { }
+  }
 }
 
 export async function clickByCheckboxAndDeleteWithConfirm({
